fix libm/common comparison signed/unsigned

+++ src/lib/libm/common/s_llround.c
@@ -51,7 +51,7 @@ llround(double x)
           result = msw >> (20 - exponent_less_1023);
         }
     }
-  else if (exponent_less_1023 < (8 * sizeof (long long int)) - 1)
+  else if ((size_t)exponent_less_1023 < (8 * sizeof (long long int)) - 1)
     {
       /* 64bit longlong: exponent_less_1023 in [20,62] */
       if (exponent_less_1023 >= 52)
@@ -71,7 +71,7 @@ llround(double x)
 	  /* 64bit longlong: shift amt in [0,31] */
           result = ((long long int) msw << (exponent_less_1023 - 20))
 		    /* ***64bit longlong: shift amt in [32,1] */
-                    | SAFE_RIGHT_SHIFT (tmp, (52 - exponent_less_1023));
+                    | SAFE_RIGHT_SHIFT (tmp, (unsigned int)(52 - exponent_less_1023));
         }
     }
   else
